Moved commit and restore from vm to session
